Fixed floating point error in pixelPush bedRockCount inf loop condition
In pixelPush, when pushDistance is calculated, on mobile, sometimes the
target.x/y positions are fractional. Therefore the final pushDistance
after it has been decremented to 0 during orthPixelPush call, must be
truncated with Math.trunc() when checking for infinite loops and
bedRockCount is incremented up to 20 max before abandoning collision
processing.